Recently discussed updates
Reads a file that doesn't exist yet in the main branch 'whi.csv' to a dataframe and drops all the columns other than the happiness score. Integrates Thomas' updates that include the map projection but I have no idea how it's going to work when it's off of colab, especially since I didn't modify the last argument in line 113. Will continue working on this and if this allows me to see a displayed version of this branch, will debug.